Fix double hex to decimal conversion (#3267)
In the previous commit with hash a3cf5c that aimed to address the issue
where sfputil incorrectly interpreted page numbers as decimal instead of
hexadecimal, there was an inadvertent double conversion from hexadecimal
to decimal. For instance, inputting 11 resulted in conversion to 17 and
then further to 23. To rectify this, the second conversion would be
removed.

A related ut has also been added.

Signed-off-by: Yuanzhe, Liu <yualiu@nvidia.com>